razer-mute-link/requirements.txt
pywinusb==0.4.2
pycaw==20240210
comtypes>=1.1.14
pystray==0.19.5
pillow>=10.0.0
